towards warning-free -wconversion

+++ test/core/test_random.cpp
@@ -6,10 +6,10 @@
 
 BOOST_AUTO_TEST_CASE(test_random)
 {
-        const size_t tests = 1024;
-        const size_t test_size = 1024;
+        const int32_t tests = 1024;
+        const int32_t test_size = 1024;
 
-        for (size_t t = 0; t < tests; t ++)
+        for (int32_t t = 0; t < tests; t ++)
         {
                 const int32_t min = 17 + t;
                 const int32_t max = min + t * 25 + 4;
@@ -18,7 +18,7 @@ BOOST_AUTO_TEST_CASE(test_random)
                 math::random_t<int32_t> rgen(min, max);
 
                 // check generator
-                for (size_t tt = 0; tt < test_size; tt ++)
+                for (int32_t tt = 0; tt < test_size; tt ++)
                 {
                         const int32_t v = rgen();
                         BOOST_CHECK_GE(v, min);
